returned location publisher to launch
For now, rather than creating a package, we
'll just have to launch this in a separate terminal when we launch the adventurebot.
After spending a fair amount of time spinning my wheels on package creation, cleaning this up
is a lower priority than completing the transition from robot_pose_ekf to robot_localization.