Switching to map
Removed all manual origin tranformations. From here on out, we will relie on
transformations between odometry and the map frame to get current position
info. All this should be handled by the recently created map filter. I'm
expecting some serious debugging to be on the horizon, but we'll figure out
a way to make it though.